+++ src/main/java/com/tierlist/tierlist/category/adapter/in/web/CategoryFavoriteController.java
@@ -1,4 +1,4 @@
-package com.tierlist.tierlist.category.adaptor.in.web;
+package com.tierlist.tierlist.category.adapter.in.web;
 
 import com.tierlist.tierlist.category.application.port.in.service.CategoryFavoriteUseCase;
 import lombok.RequiredArgsConstructor;
